third_party/WebKit/LayoutTests/fast/dom/Window/Location/location-override-valueOf-on-proto.html

<html>
<head>
    <script src="../../../../resources/js-test.js"></script>
    <script src="resources/location-tests-functions.js"></script>
</head>
<body>
<script>
    window.location.__proto__.valueOf = function() { return "haxored"; }

    var result = normalizeURL(String(window.location));
    var correctValue = normalizeURL(document.URL);
    shouldBe("result", "correctValue");
</script>
</body>
</html>
